+++ saleor/graphql/account/mutations/permission_group/__init__.py
@@ -1,0 +1,9 @@
+from .permission_group_create import PermissionGroupCreate
+from .permission_group_delete import PermissionGroupDelete
+from .permission_group_update import PermissionGroupUpdate
+
+__all__ = [
+    "PermissionGroupCreate",
+    "PermissionGroupDelete",
+    "PermissionGroupUpdate",
+]
